requirements.txt
asgiref==3.10.0
Django==4.2.25
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
